+++ packages/myst/lib/src/services/services.dart
@@ -1,1 +1,2 @@
-export './file_services.dart';
+export 'file_services.dart';
+export 'generate_file_helper.dart';